Fix logout logic
The default who.ini configuration provided meant that the friendlyform
plugin was masking all the logic, and it resulted in the user being
asked to sign in again in the Identity Provider while still being left
logged in the application (ie cookies in the browser).

Once ckanext-saml2 was able to execute the logout logic, there were
errors in there (in fact this code likely never worked). Updated it to
use the entity id from the configuration and to actually delete the auth
cookies.